Check for session in sign up page
